move headless AppConfig type and associated types to fiori gen shared module
