+++ .changeset/cyan-timers-jog.md
@@ -1,0 +1,8 @@
+---
+'@sap-ux/deploy-config-sub-generator': patch
+'@sap-ux/fiori-app-sub-generator': patch
+'@sap-ux/fiori-generator-shared': patch
+'@sap-ux/cap-config-writer': patch
+---
+
+move headless AppConfig type and associated types to fiori gen shared module
