+++ packages/cap-config-writer/src/index.ts
@@ -1,6 +1,7 @@
 import { checkCdsUi5PluginEnabled } from '@sap-ux/project-access';
 export { checkCdsUi5PluginEnabled };
 export { enableCdsUi5Plugin } from './cap-config';
-export type { CapService, CapRuntime } from './cap-config/types';
+export type { CapService } from './cap-config/types';
+export type { CapRuntime } from '@sap-ux/fiori-generator-shared';
 export type { CapServiceCdsInfo, CapProjectSettings } from './cap-config/types';
 export * from './cap-writer';
