+++ packages/deploy-config-sub-generator/tsconfig.json
@@ -31,9 +31,6 @@
       "path": "../deploy-config-generator-shared"
     },
     {
-      "path": "../fiori-app-sub-generator"
-    },
-    {
       "path": "../fiori-generator-shared"
     },
     {
