+++ packages/fiori-app-sub-generator/src/types/external.ts
@@ -1,8 +1,7 @@
 import type { IPrompt as Step } from '@sap-devx/yeoman-ui-types';
-import type { Annotations } from '@sap-ux/axios-extension';
 import { TemplateType as FETemplateType } from '@sap-ux/fiori-elements-writer';
 import { TemplateType as FFTemplateType } from '@sap-ux/fiori-freestyle-writer';
-import type { DeployConfig, FLPConfig } from '@sap-ux/fiori-generator-shared';
+import { AppConfig, type Floorplan, FloorplanFE, FloorplanFF } from '@sap-ux/fiori-generator-shared';
 import type { CapRuntime, EntityRelatedAnswers } from '@sap-ux/odata-service-inquirer';
 import { OdataVersion } from '@sap-ux/odata-service-inquirer';
 import {
@@ -11,12 +10,9 @@ import {
 } from '@sap-ux/ui5-application-inquirer';
 import type { Answers } from 'inquirer';
 import { LEGACY_CAP_TYPE_JAVA, LEGACY_CAP_TYPE_NODE } from './constants';
-import { type ALPOptions, type Project, type Service, type Floorplan, FloorplanFE, FloorplanFF } from './state';
+import { type ALPOptions, type Project, type Service } from './state';
 
 export { Floorplan, FloorplanFE, FloorplanFF };
-
-// Used in external interfaces to define floorplans using a simple meaningful string key
-export type FloorplanKey = keyof typeof FloorplanFE | keyof typeof FloorplanFF;
 
 type FloorplanAttributesType = {
     [K in Floorplan]: {
@@ -56,59 +52,6 @@ export const FloorplanAttributes: FloorplanAttributesType = {
         templateType: FFTemplateType.Basic
     }
 };
-
-/**
- * Defines the external interface used to generate in headless mode (no prompts)
- * This is a deliberate re-definition of internal interfaces to avoid consumers having
- * to update when internal interfaces are changed
- * NOTE: Any breaking changes to this interface require a version bump
- */
-export interface AppConfig {
-    readonly version: string; // The interface version
-    readonly floorplan: FloorplanKey;
-    project: {
-        readonly name: string;
-        targetFolder?: string; // Current working directory will be used if not provided
-        readonly namespace?: string;
-        readonly title?: string;
-        readonly description?: string;
-        readonly ui5Theme?: string;
-        readonly ui5Version?: string;
-        readonly localUI5Version?: string;
-        readonly sapux?: boolean;
-        readonly skipAnnotations?: boolean;
-        readonly enableCodeAssist?: boolean;
-        readonly enableEslint?: boolean;
-        readonly enableTypeScript?: boolean;
-    };
-    service?: {
-        readonly host?: string;
-        readonly servicePath?: string;
-        readonly client?: string;
-        readonly scp?: boolean; // If available key store entry must be available or provided at app runtime
-        readonly destination?: string;
-        readonly destinationInstance?: string;
-        readonly edmx?: string;
-        readonly annotations?: Annotations | Annotations[];
-        readonly capService?: {
-            readonly projectPath: string;
-            readonly serviceName: string;
-            readonly serviceCdsPath: string;
-            readonly capType?: CapRuntime;
-            readonly appPath?: string; // Alternative app path
-        };
-        readonly apiHubApiKey?: string; // Non-enterprise support only currently
-    };
-    deployConfig?: DeployConfig;
-    flpConfig?: FLPConfig;
-    /**
-     * Adds telemetry data when passed to generator `@sap/generator-fiori:headless`
-     */
-    telemetryData?: {
-        generationSourceName?: string;
-        generationSourceVersion?: string;
-    };
-}
 
 /**
  * Defines the entity config property of the external app config interface used to generate in headless mode (no prompts)
