+++ packages/fiori-generator-shared/package.json
@@ -45,6 +45,7 @@
         "@types/semver": "7.5.2",
         "@types/vscode": "1.73.1",
         "@types/yeoman-environment": "2.10.11",
+        "@sap-ux/axios-extension": "workspace:*",
         "@sap-ux/logger": "workspace:*"
     },
     "engines": {
