+++ packages/fiori-generator-shared/src/types/cap.ts
@@ -1,0 +1,1 @@
+export type CapRuntime = 'Node.js' | 'Java';
