+++ packages/fiori-generator-shared/src/types/index.ts
@@ -1,3 +1,4 @@
 export * from './app-gen';
+export * from './cap';
 export * from './environment';
 export * from './headless';
